+++ fidget-raster/src/lib.rs
@@ -475,16 +475,12 @@ pub struct BadPixelCount {
     pub height: u32,
 }
 
-/// Error type for render operations builder
+/// Error type for render operations
 #[derive(thiserror::Error, Debug, PartialEq)]
 pub enum RenderError {
-    /// Missing variable in the [`ShapeVars`](fidget_core::shape::ShapeVars) map
+    /// Missing variable in the [`ShapeVars`] map
     #[error(transparent)]
     MissingVar(#[from] fidget_core::shape::MissingVar),
-
-    /// Render was cancelled
-    #[error("render was cancelled")]
-    Cancelled,
 }
 
 #[cfg(test)]