+++ fidget/tests/pixel_render.rs
@@ -45,6 +45,7 @@ impl Cfg {
         };
         let out = cfg
             .run_with_vars(shape, &self.vars)
+            .expect("rendering should not fail")
             .expect("rendering should not be cancelled");
         let mut img_str = String::new();
         for (i, b) in out.iter().enumerate() {
@@ -379,7 +380,10 @@ fn check_neg_infinity<F: Function + MathFunction + RenderHints>() {
         threads: None,
         ..Default::default()
     };
-    let out = cfg.run(shape).unwrap();
+    let out = cfg
+        .run(shape)
+        .expect("rendering should not fail")
+        .expect("rendering should not be cancelled");
     assert!(out.into_iter().all(|i| i.inside()));
 }